LayoutTests/imported/w3c/web-platform-tests/html/browsers/the-window-object/garbage-collection-and-browsing-contexts/discard_iframe_history_2-expected.txt


FAIL Removing iframe from document via innerHTML removes it from history assert_equals: History length before iframe removal expected 2 but got 1
